+++ routes/board/views/components/index.js
@@ -1,0 +1,9 @@
+const { reply } = require('./reply');
+const { form } = require('./form');
+const { op } = require('./op');
+
+module.exports = {
+  reply,
+  form,
+  op
+};