+++ routes/board/views/index.js
@@ -1,0 +1,9 @@
+const { thread } = require('./thread');
+const { list } = require('./list');
+const { nu } = require('./nu');
+
+module.exports = {
+  thread,
+  list,
+  nu
+};