+++ routes/index.js
@@ -1,7 +1,9 @@
-const api = require('./api');
 const proxy = require('./proxy');
+const board = require('./board');
+const api = require('./api');
 
 module.exports = {
   api,
-  proxy
+  proxy,
+  board
 };